+++ services/deduplicate/index.ts
@@ -1,4 +1,7 @@
 import { select, dispatch } from '@wordpress/data';
+import apiFetch from '@wordpress/api-fetch';
+import { addQueryArgs } from '@wordpress/url';
+import type { WP_REST_API_Posts as WpRestApiPosts } from 'wp-types'; // eslint-disable-line camelcase
 
 const usedIds = new Map();
 const curatedIds = new Map();
@@ -81,7 +84,7 @@ const getQueryBlocks = (blocks: Block[], out: Block[]) => {
  * This is the main function to update all pinned posts. Call it whenever a pinned post
  * changes or the query settings change.
  */
-export function mainDedupe() {
+export async function mainDedupe() {
   if (running) {
     // Only one run at a time, but mark that another run has been requested.
     redo = true;
@@ -96,29 +99,29 @@ export function mainDedupe() {
   const blocks: Block[] = select('core/block-editor').getBlocks();
   const {
     wp_curate_deduplication: wpCurateDeduplication = true,
+    wp_curate_unique_pinned_posts: wpCurateUniquePinnedPosts = false,
     // @ts-ignore
   } = select('core/editor').getEditedPostAttribute('meta') || {};
 
   const queryBlocks: Block[] = [];
   // Loop through all blocks and find all query blocks.
   getQueryBlocks(blocks, queryBlocks);
-  // Loop through all query blocks and find all pinned posts. Add them to the list of used ids.
-  // Note: This means that a pinned post will be used only once on the page. It cannot be backfilled
-  // earlier on the page. The front end does not currently do this, so it will be commented out
-  // here to match. We should probably add a setting to allow this in both places.
 
-  // queryBlocks.forEach((queryBlock) => {
-  //   const { attributes } = queryBlock;
-  //   const { posts } = attributes;
-  //   posts?.forEach((post) => {
-  //     if (post) {
-  //       deduplicate(post);
-  //     }
-  //   });
-  // });
+  /**
+   * This block of code is responsible for enforcing the unique pinned posts setting in the editor.
+   */
+  if (wpCurateUniquePinnedPosts) {
+    queryBlocks.forEach((queryBlock) => {
+      queryBlock?.attributes?.posts?.forEach((post) => {
+        if (post) {
+          deduplicate(post);
+        }
+      });
+    });
+  }
 
   // Loop through all query blocks and set backfilled posts in the open slots.
-  queryBlocks.forEach((queryBlock) => {
+  queryBlocks.forEach(async (queryBlock) => {
     const { attributes } = queryBlock;
     const {
       backfillPosts = null,
@@ -130,6 +133,26 @@ export function mainDedupe() {
     if (!backfillPosts) {
       return;
     }
+
+    const postsToInclude = posts.filter((id) => id !== null).join(',');
+    let validPosts: Number[] = [];
+
+    if (postsToInclude.length > 0) {
+      validPosts = await apiFetch({
+        path: addQueryArgs(
+          '/wp/v2/posts',
+          {
+            offset: 0,
+            orderby: 'include',
+            per_page: posts.length,
+            type: 'post',
+            include: postsToInclude,
+            _locale: 'user',
+          },
+        ),
+      }).then((response) => (response as any as WpRestApiPosts).map((post) => post.id));
+    }
+
     const postTypeString = postTypes.join(',');
     let postIndex = 0;
 
@@ -137,7 +160,9 @@ export function mainDedupe() {
     const allPostIds: Array<number | undefined> = [];
 
     // New array to hold the pinned posts in the order they should be.
-    const manualPostIdArray: Array<number | null> = posts;
+    const manualPostIdArray: Array<number | null> = posts.map(
+      (post) => validPosts.includes(post) ? post : null, // eslint-disable-line no-confusing-arrow
+    );
 
     // Remove any pinned posts from the backfilled posts list.
     const filteredPosts = backfillPosts.filter((post) => !manualPostIdArray.includes(post));